# Improve memory management, add benchmark tests