## Goals

1. Add benchmark tests to evaluate library performance.
2. Refactor context memory management to enable thread-safe protection across branching additions without duplicating existing data at each addition.
3. Refactor error memory management to enable thread-safe protection across error context additions without duplicating error objects at each addition.

## Benchmarks

Some notable insights on the following benchmarks:
* The overall goal of this change is to ensure no single action gets out of hand.  Additional performance is nice, and can be improved still in the future, but maximizing performance across the board is not the current goal.
* The `StaticK*` benchmarks are technically the most accurate representations of real world usage, since they contain a minimum overlap across keys.  However these tests will, in all cases, stack data much deeper (thousands of nodes) than is realistically expected of a production environment (tens of nodes).
* While performance sees reasonable improvement in the general case, solving the `InErr()` timeouts is the biggest win.

### Before Changes
```
BenchmarkAdd_singleConstKConstV-10               1376434               789.9 ns/op           992 B/op          8 allocs/op
BenchmarkAdd_singleStaticKStaticV-10               10000            323316 ns/op          338917 B/op         22 allocs/op
BenchmarkAdd_singleConstKStaticV-10              1619610               743.1 ns/op          1016 B/op         10 allocs/op
BenchmarkAdd_singleStaticKConstV-10                10000            265893 ns/op          338878 B/op         19 allocs/op
BenchmarkAdd_singleConstKRandV-10                1489209               811.2 ns/op          1048 B/op         11 allocs/op
BenchmarkAdd_singleRandKConstV-10                  10000            168960 ns/op          238314 B/op         13 allocs/op
BenchmarkAdd_singleRandKRandV-10                   10000            192365 ns/op          238371 B/op         16 allocs/op
BenchmarkAdd_multConstKConstV-10                 1760774              1026 ns/op            1008 B/op          9 allocs/op
BenchmarkAdd_multStaticKStaticV-10                 10000            395954 ns/op          470004 B/op         29 allocs/op
BenchmarkAdd_multConstKStaticV-10                1000000              1442 ns/op            1056 B/op         14 allocs/op
BenchmarkAdd_multStaticKConstV-10                  10000            694954 ns/op          671398 B/op         37 allocs/op
BenchmarkAdd_multConstKRandV-10                   781983              1384 ns/op            1120 B/op         15 allocs/op
BenchmarkAdd_multRandKConstV-10                    10000            391174 ns/op          469039 B/op         20 allocs/op
BenchmarkAdd_multRandKRandV-10                     10000            433996 ns/op          469144 B/op         26 allocs/op
BenchmarkAddMap_constKConstV-10                  1186197              1565 ns/op            1136 B/op         14 allocs/op
BenchmarkAddMap_staticKStaticV-10                  10000            384590 ns/op          470112 B/op         32 allocs/op
BenchmarkAddMap_constKStaticV-10                  999336              1692 ns/op            1152 B/op         18 allocs/op
BenchmarkAddMap_staticKConstV-10                   10000            686717 ns/op          671478 B/op         40 allocs/op
BenchmarkAddMap_constKRandV-10                   1000000              1985 ns/op            1216 B/op         18 allocs/op
BenchmarkAddMap_randKConstV-10                     10000            371346 ns/op          469125 B/op         23 allocs/op
BenchmarkAddMap_randKRandV-10                      10000            393378 ns/op          469248 B/op         29 allocs/op
BenchmarkIn_const-10                             1564428               688.0 ns/op           992 B/op          8 allocs/op
BenchmarkIn_static-10                              10000            286547 ns/op          338913 B/op         22 allocs/op
BenchmarkIn_rand-10                                10000            188766 ns/op          238371 B/op         16 allocs/op
BenchmarkWith_singleConstKConstV-10              6823659               241.0 ns/op           448 B/op          4 allocs/op
BenchmarkWith_singleStaticKStaticV-10            3834045               336.9 ns/op           496 B/op          9 allocs/op
BenchmarkWith_singleConstKStaticV-10             4493137               262.1 ns/op           472 B/op          6 allocs/op
BenchmarkWith_singleStaticKConstV-10             4624930               261.4 ns/op           472 B/op          6 allocs/op
BenchmarkWith_singleConstKRandV-10               4286445               299.4 ns/op           503 B/op          7 allocs/op
BenchmarkWith_singleRandKConstV-10               4351930               289.1 ns/op           503 B/op          7 allocs/op
BenchmarkWith_singleRandKRandV-10                3232315               371.3 ns/op           559 B/op         10 allocs/op
BenchmarkWith_multConstKConstV-10                5094906               479.2 ns/op           464 B/op          5 allocs/op
BenchmarkWith_multStaticKStaticV-10              2126823               838.3 ns/op           560 B/op         16 allocs/op
BenchmarkWith_multConstKStaticV-10               2991151               482.1 ns/op           512 B/op         10 allocs/op
BenchmarkWith_multStaticKConstV-10               2782454               593.2 ns/op           512 B/op         10 allocs/op
BenchmarkWith_multConstKRandV-10                 2749723               694.1 ns/op           575 B/op         11 allocs/op
BenchmarkWith_multRandKConstV-10                 2379142               462.6 ns/op           575 B/op         11 allocs/op
BenchmarkWith_multRandKRandV-10                  1314402               874.5 ns/op           687 B/op         17 allocs/op
BenchmarkWith_chainConstKConstV-10               3051607               378.1 ns/op           896 B/op          8 allocs/op
BenchmarkWith_chainStaticKStaticV-10             1849600               692.0 ns/op           992 B/op         19 allocs/op
BenchmarkWith_chainConstKStaticV-10              2259436               535.7 ns/op           944 B/op         13 allocs/op
BenchmarkWith_chainStaticKConstV-10              2206014               525.2 ns/op           944 B/op         13 allocs/op
BenchmarkWith_chainConstKRandV-10                2176041               596.0 ns/op          1007 B/op         14 allocs/op
BenchmarkWith_chainRandKConstV-10                2150919               600.0 ns/op          1007 B/op         14 allocs/op
BenchmarkWith_chainRandKRandV-10                 1626723               759.9 ns/op          1119 B/op         20 allocs/op
BenchmarkWithMap_constKConstV-10                 7203932               173.6 ns/op           432 B/op          3 allocs/op
BenchmarkWithMap_staticKStaticV-10               5727504               216.7 ns/op           463 B/op          6 allocs/op
BenchmarkWithMap_constKStaticV-10                6696054               166.6 ns/op           448 B/op          4 allocs/op
BenchmarkWithMap_staticKConstV-10                5697630               194.0 ns/op           447 B/op          4 allocs/op
BenchmarkWithMap_constKRandV-10                  7057366               172.0 ns/op           448 B/op          5 allocs/op
BenchmarkWithMap_randKConstV-10                  5308044               213.1 ns/op           479 B/op          5 allocs/op
BenchmarkWithMap_randKRandV-10                   5014779               228.9 ns/op           495 B/op          7 allocs/op
BenchmarkWithClues_constKConstV-10               2031816               649.2 ns/op          1088 B/op          9 allocs/op
BenchmarkWithClues_staticKStaticV-10               10000            249326 ns/op          339016 B/op         23 allocs/op
BenchmarkWithClues_constKStaticV-10              1402378               928.2 ns/op          1112 B/op         11 allocs/op
BenchmarkWithClues_staticKConstV-10                10000            237654 ns/op          338967 B/op         20 allocs/op
BenchmarkWithClues_constKRandV-10                1493005               817.6 ns/op          1144 B/op         12 allocs/op
BenchmarkWithClues_randKConstV-10                  10000            190073 ns/op          238414 B/op         14 allocs/op
BenchmarkWithClues_randKRandV-10                   10000            161378 ns/op          238473 B/op         17 allocs/op
BenchmarkInErr_const-10                            10000           1432733 ns/op         1680719 B/op      10007 allocs/op
BenchmarkInErr_static-10                        fSIGQUIT: quit
*** Test killed with quit: ran too long (11m0s).
exit status 2
FAIL    github.com/alcionai/clues       660.059s
```

### After Changes
```
BenchmarkAdd_singleConstKConstV-10               1928366               824.2 ns/op           784 B/op          9 allocs/op
BenchmarkAdd_singleStaticKStaticV-10             1628745               787.6 ns/op           832 B/op         14 allocs/op
BenchmarkAdd_singleConstKStaticV-10              1658325               775.3 ns/op           808 B/op         11 allocs/op
BenchmarkAdd_singleStaticKConstV-10              1908662               553.3 ns/op           808 B/op         11 allocs/op
BenchmarkAdd_singleConstKRandV-10                1863534               568.9 ns/op           840 B/op         12 allocs/op
BenchmarkAdd_singleRandKConstV-10                1910954               563.9 ns/op           840 B/op         12 allocs/op
BenchmarkAdd_singleRandKRandV-10                 1752163               712.1 ns/op           896 B/op         15 allocs/op
BenchmarkAdd_multConstKConstV-10                 1989670               559.1 ns/op           800 B/op         10 allocs/op
BenchmarkAdd_multStaticKStaticV-10               1242932               993.5 ns/op           896 B/op         21 allocs/op
BenchmarkAdd_multConstKStaticV-10                1000000              1250 ns/op             848 B/op         16 allocs/op
BenchmarkAdd_multStaticKConstV-10                1616938              1089 ns/op             848 B/op         15 allocs/op
BenchmarkAdd_multConstKRandV-10                  1265910              1033 ns/op             912 B/op         16 allocs/op
BenchmarkAdd_multRandKConstV-10                  1000000              1027 ns/op             912 B/op         16 allocs/op
BenchmarkAdd_multRandKRandV-10                   1000000              1219 ns/op            1024 B/op         22 allocs/op
BenchmarkAddMap_constKConstV-10                  1326388               993.4 ns/op           928 B/op         15 allocs/op
BenchmarkAddMap_staticKStaticV-10                1000000              1126 ns/op            1008 B/op         24 allocs/op
BenchmarkAddMap_constKStaticV-10                 1000000              1036 ns/op             944 B/op         18 allocs/op
BenchmarkAddMap_staticKConstV-10                 1000000              1409 ns/op             944 B/op         18 allocs/op
BenchmarkAddMap_constKRandV-10                   1000000              1088 ns/op            1008 B/op         19 allocs/op
BenchmarkAddMap_randKConstV-10                   1000000              1168 ns/op            1008 B/op         19 allocs/op
BenchmarkAddMap_randKRandV-10                    1054890              1528 ns/op            1136 B/op         25 allocs/op
BenchmarkIn_const-10                             2275856               482.1 ns/op           784 B/op          9 allocs/op
BenchmarkIn_static-10                            1743090               628.2 ns/op           832 B/op         14 allocs/op
BenchmarkIn_rand-10                              1775120               694.1 ns/op           895 B/op         15 allocs/op
BenchmarkWith_singleConstKConstV-10              3284148               343.2 ns/op           704 B/op          6 allocs/op
BenchmarkWith_singleStaticKStaticV-10            2095527               531.9 ns/op           752 B/op         11 allocs/op
BenchmarkWith_singleConstKStaticV-10             2479923               454.5 ns/op           728 B/op          8 allocs/op
BenchmarkWith_singleStaticKConstV-10             2787708               463.9 ns/op           728 B/op          8 allocs/op
BenchmarkWith_singleConstKRandV-10               2391356               481.7 ns/op           760 B/op          9 allocs/op
BenchmarkWith_singleRandKConstV-10               2299364               476.4 ns/op           760 B/op          9 allocs/op
BenchmarkWith_singleRandKRandV-10                2098438               565.7 ns/op           815 B/op         12 allocs/op
BenchmarkWith_multConstKConstV-10                2728303               735.6 ns/op           720 B/op          7 allocs/op
BenchmarkWith_multStaticKStaticV-10              1412894              1072 ns/op             816 B/op         18 allocs/op
BenchmarkWith_multConstKStaticV-10               2025582               714.1 ns/op           768 B/op         12 allocs/op
BenchmarkWith_multStaticKConstV-10               1776141               629.9 ns/op           768 B/op         12 allocs/op
BenchmarkWith_multConstKRandV-10                 1747958              1011 ns/op             832 B/op         13 allocs/op
BenchmarkWith_multRandKConstV-10                 1722884               788.2 ns/op           832 B/op         13 allocs/op
BenchmarkWith_multRandKRandV-10                  1000000              1384 ns/op             944 B/op         19 allocs/op
BenchmarkWith_chainConstKConstV-10               1608531               709.4 ns/op          1408 B/op         12 allocs/op
BenchmarkWith_chainStaticKStaticV-10             1000000              1084 ns/op            1504 B/op         23 allocs/op
BenchmarkWith_chainConstKStaticV-10              1390452               940.4 ns/op          1456 B/op         17 allocs/op
BenchmarkWith_chainStaticKConstV-10              1229997              1263 ns/op            1456 B/op         17 allocs/op
BenchmarkWith_chainConstKRandV-10                1193283               966.7 ns/op          1520 B/op         18 allocs/op
BenchmarkWith_chainRandKConstV-10                1258813               950.1 ns/op          1520 B/op         18 allocs/op
BenchmarkWith_chainRandKRandV-10                 1000000              1235 ns/op            1632 B/op         24 allocs/op
BenchmarkWithMap_constKConstV-10                 5340348               251.8 ns/op           352 B/op          3 allocs/op
BenchmarkWithMap_staticKStaticV-10               4485530               296.7 ns/op           383 B/op          6 allocs/op
BenchmarkWithMap_constKStaticV-10                4862736               245.5 ns/op           367 B/op          4 allocs/op
BenchmarkWithMap_staticKConstV-10                4787144               270.4 ns/op           367 B/op          4 allocs/op
BenchmarkWithMap_constKRandV-10                  4957141               248.2 ns/op           368 B/op          5 allocs/op
BenchmarkWithMap_randKConstV-10                  4572518               287.1 ns/op           399 B/op          5 allocs/op
BenchmarkWithMap_randKRandV-10                   4263903               312.9 ns/op           415 B/op          7 allocs/op
BenchmarkWithClues_constKConstV-10                 10000            276624 ns/op            1472 B/op         14 allocs/op
BenchmarkWithClues_staticKStaticV-10               10000           1197163 ns/op         1052102 B/op        142 allocs/op
BenchmarkWithClues_constKStaticV-10                10000            277806 ns/op            1496 B/op         16 allocs/op
BenchmarkWithClues_staticKConstV-10                10000           1224210 ns/op         1051874 B/op        138 allocs/op
BenchmarkWithClues_constKRandV-10                  10000            276510 ns/op            1529 B/op         17 allocs/op
BenchmarkWithClues_randKConstV-10                  10000            919043 ns/op          739279 B/op         89 allocs/op
BenchmarkWithClues_randKRandV-10                   10000            852452 ns/op          739337 B/op         92 allocs/op
BenchmarkInErr_const-10                            10000            248625 ns/op            1424 B/op         11 allocs/op
BenchmarkInErr_static-10                           10000           1247076 ns/op         1428164 B/op        246 allocs/op
BenchmarkInErr_rand-10                             10000            957139 ns/op         1003028 B/op        158 allocs/op
PASS
ok      github.com/alcionai/clues       182.191s
```